B578A first homework assignment due 2014-02-04
1. Use the GEOmetabd package to find all HCV gene expression data using
the Illumina platform submitted by an investigator at Yale. This should
be done with a single query, showing the title, the GSE accession
number, the GPL accession number and the manufacturer and the
description of the platform used.

2. Reproduce your above query using the data.table package. Again, try
to use a single line of code. [Hint: You first need to convert all db
tables to data.table tables].